+++ instrumentation/google.golang.org/grpc/otelgrpc/stats_handler.go
@@ -44,6 +44,9 @@ type serverHandler struct {
 	outMsg   rpcconv.ServerResponsesPerRPC
 }
 
+// defaultByteHistogramBounds defines the default By unit explicit bucket bounds: [0, 50MB].
+var defaultByteHistogramBounds = []float64{100_000, 500_000, 1000_000, 1500_000, 2000_000, 2500_000, 3000_000, 3500_000, 4000_000, 8000_000, 10000_000, 2000_0000, 30000_000, 40000_000, 50000_000}
+
 // NewServerHandler creates a stats.Handler for a gRPC server.
 func NewServerHandler(opts ...Option) stats.Handler {
 	c := newConfig(opts)
@@ -66,12 +69,12 @@ func NewServerHandler(opts ...Option) stats.Handler {
 		otel.Handle(err)
 	}
 
-	h.inSize, err = rpcconv.NewServerRequestSize(meter)
+	h.inSize, err = rpcconv.NewServerRequestSize(meter, metric.WithExplicitBucketBoundaries(defaultByteHistogramBounds...))
 	if err != nil {
 		otel.Handle(err)
 	}
 
-	h.outSize, err = rpcconv.NewServerResponseSize(meter)
+	h.outSize, err = rpcconv.NewServerResponseSize(meter, metric.WithExplicitBucketBoundaries(defaultByteHistogramBounds...))
 	if err != nil {
 		otel.Handle(err)
 	}